Migrate Filter Code Appropriately
