+++ src/Traits/Core/Filters/HandlesFilterTraits.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Rappasoft\LaravelLivewireTables\Traits\Core\Filters;
+
+trait HandlesFilterTraits
+{
+    use ManagesFilters,
+        HasFilterGenericData,
+        HasFilterMenuStyling,
+        HasFilterPillsStyling,
+        HasFilterQueryString,
+        HasFiltersVisibility;
+}
