+++ src/Traits/Core/Filters/HasFilterGenericData.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace Rappasoft\LaravelLivewireTables\Traits\Core\Filters;
+
+use Rappasoft\LaravelLivewireTables\DataTransferObjects\FilterGenericData;
+use Livewire\Attributes\Computed;
+
+trait HasFilterGenericData
+{
+    public array $filterGenericData = [];
+
+    public function generateFilterGenericData(): array
+    {
+        return (new FilterGenericData($this->getTableName(), $this->getFilterLayout(), $this->isTailwind(), $this->isBootstrap4(), $this->isBootstrap5()))->toArray();
+    }
+
+    public function setFilterGenericData(array $filterGenericData = []): void
+    {
+        $this->filterGenericData = $filterGenericData;
+    }
+
+    public function hasFilterGenericData(): bool
+    {
+        return ! empty($this->filterGenericData);
+    }
+
+    #[Computed]
+    public function getFilterGenericData(): array
+    {
+        if (! $this->hasFilterGenericData()) {
+            $this->setFilterGenericData($this->generateFilterGenericData());
+        }
+
+        return $this->filterGenericData;
+    }
+}
