+++ src/Traits/Core/Filters/HasFilterPillsStyling.php
@@ -1,0 +1,63 @@
+<?php
+
+namespace Rappasoft\LaravelLivewireTables\Traits\Core\Filters;
+
+use Livewire\Attributes\{Computed, Locked};
+use Rappasoft\LaravelLivewireTables\Views\Filter;
+
+trait HasFilterPillsStyling
+{
+    #[Locked]
+    public bool $filterPillsStatus = true;
+
+    public function setFilterPillsStatus(bool $status): self
+    {
+        $this->filterPillsStatus = $status;
+
+        return $this;
+    }
+
+    public function setFilterPillsEnabled(): self
+    {
+        $this->setFilterPillsStatus(true);
+
+        return $this;
+    }
+
+    public function setFilterPillsDisabled(): self
+    {
+        $this->setFilterPillsStatus(false);
+
+        return $this;
+    }
+
+    #[Computed]
+    public function showFilterPillsSection(): bool
+    {
+        return $this->filtersAreEnabled() && $this->filterPillsAreEnabled() && $this->hasAppliedVisibleFiltersForPills();
+    }
+
+    public function getFilterPillsStatus(): bool
+    {
+        return $this->filterPillsStatus;
+    }
+
+    public function filterPillsAreEnabled(): bool
+    {
+        return $this->getFilterPillsStatus() === true;
+    }
+
+    public function filterPillsAreDisabled(): bool
+    {
+        return $this->getFilterPillsStatus() === false;
+    }
+
+    public function hasAppliedVisibleFiltersForPills(): bool
+    {
+        return collect($this->getAppliedFiltersWithValues())
+            ->map(fn ($_item, $key) => $this->getFilterByKey($key))
+            ->reject(fn (Filter $filter) => $filter->isHiddenFromPills())
+            ->count() > 0;
+    }
+
+}
