+++ src/Traits/Core/Filters/ManagesFilters.php
@@ -1,0 +1,106 @@
+<?php
+
+namespace Rappasoft\LaravelLivewireTables\Traits\Core\Filters;
+
+use Illuminate\Database\Eloquent\Builder;
+use Rappasoft\LaravelLivewireTables\Events\FilterApplied;
+
+trait ManagesFilters
+{        
+
+    // Set in JS
+    public array $filterComponents = [];
+
+    // Set in Frontend
+    public array $appliedFilters = [];
+
+
+        /**
+     * Sets Filter Default Values
+     */
+    public function mountManagesFilters(): void
+    {
+        $this->restoreFilterValues();
+
+        foreach ($this->getFilters() as $filter) {
+            if (! isset($this->appliedFilters[$filter->getKey()])) {
+                if ($filter->hasFilterDefaultValue()) {
+                    $this->setFilter($filter->getKey(), $filter->getFilterDefaultValue());
+                } else {
+                    $this->resetFilter($filter);
+                }
+            } else {
+                $this->setFilter($filter->getKey(), $this->appliedFilters[$filter->getKey()]);
+            }
+        }
+    }
+
+    public function bootedManagesFilters(): void
+    {
+        $this->setBuilder($this->builder());
+
+        foreach($this->filterComponents as $filterKey => $value)
+        {
+            $this->appliedFilters[$filterKey] = $value;
+        }
+    }
+
+    public function applyFilters(): Builder
+    {
+        if ($this->filtersAreEnabled() && $this->hasFilters() && $this->hasAppliedFiltersWithValues()) {
+            foreach ($this->getFilters() as $filter) {
+                foreach ($this->getAppliedFiltersWithValues() as $key => $value) {
+                    if ($filter->getKey() === $key && $filter->hasFilterCallback()) {
+                        // Let the filter class validate the value
+                        $value = $filter->validate($value);
+
+                        if ($value === false) {
+                            continue;
+                        }
+
+                        $this->callHook('filterApplying', ['filter' => $filter->getKey(), 'value' => $value]);
+                        $this->callTraitHook('filterApplying', ['filter' => $filter->getKey(), 'value' => $value]);
+
+                        ($filter->getFilterCallback())($this->getBuilder(), $value);
+                    }
+                }
+            }
+            $this->storeFilterValues();
+        }
+
+        return $this->getBuilder();
+    }
+
+
+    public function updatedFilterComponents(string|array|null $value, string $filterName): void
+    {
+        $this->resetComputedPage();
+
+        // Clear bulk actions on filter - if enabled
+        if ($this->getClearSelectedOnFilter()) {
+            $this->clearSelected();
+            $this->setSelectAllDisabled();
+        }
+
+        // Clear filters on empty value
+        $filter = $this->getFilterByKey($filterName);
+
+        if ($filter && $filter->isEmpty($value)) {
+            $this->callHook('filterRemoved', ['filter' => $filter->getKey()]);
+            $this->callTraitHook('filterRemoved', ['filter' => $filter->getKey()]);
+
+            $this->resetFilter($filterName);
+        } elseif ($filter) {
+            $this->callHook('filterUpdated', ['filter' => $filter->getKey(), 'value' => $value]);
+            $this->callTraitHook('filterUpdated', ['filter' => $filter->getKey(), 'value' => $value]);
+            if ($this->getEventStatusFilterApplied() && $filter->getKey() != null && $value != null) {
+                event(new FilterApplied($this->getTableName(), $filter->getKey(), $value));
+            }
+            $this->dispatch('filter-was-set', tableName: $this->getTableName(), filterKey: $filter->getKey(), value: $value);
+
+        }
+
+    }
+
+
+}
