+++ tests/Unit/Traits/Helpers/FilterHelpersTest.php
@@ -224,10 +224,12 @@ final class FilterHelpersTest extends TestCase
                 parent::configure();
             }
         };
+        $testTable->mountManagesFilters();
+
         $testTable->configure();
         $testTable->boot();
+        $testTable->bootedManagesFilters();
         $testTable->bootedComponentUtilities();
-        $testTable->bootedWithData();
         $testTable->bootedWithColumns();
         $testTable->bootedWithColumnSelect();
         $testTable->bootedWithSecondaryHeader();
